app: Add demos for STM32 G070 / G031 / G0B1 boards.
Both the G070 and G0B1 boards are fully functional up to the ping task,
i.e. the user LED blinks and "Ping!" is sent over the UART.

The G031 application builds but is unable to flash for as-yet
undetermined reasons.  It's a much smaller board than the others Hubris
currently targets (64KB flash, 8KB RAM) and may only be useful as an
minimalism exercise, but is included here anyway.