Limit arch-specific crates to the correct targets.
Ensures that cortex-m* is only included when targetting ARM, and riscv*
when targetting RISC-V.

We currently override the riscv and riscv-rt crates to pull in fixes
from upstream and for the user-mode support I added (required as there
is no way to disable interrupts from User mode).

This also adds a semihosting-riscv feature to jefe as "semihosting"
depends on the cortex-m-semihosting crate, there may be a better way to
resolve this.